Update cql3handling.py
I moved four existing "terminal productions"  and added three new ones at lines: 174 - 180
the new ones are: <closedComment>, <closedTComment>, <unclosedTComment>

The problem was that the existing 4 regex rule or validator was too down in the set and never caught anything.
Unfortunately I couldn't investigate what's the purpose of the "JUNK " rule. 

I did several manual tests in cqlsh and this is a working fix.