Update cqlshmain.py
this is the seconds part of the fix for CASSANDRA-17667.
the strip_comment_blocks function is not useful after putting the regex rules in the correct order in cql3handling file.
in fact strip_comment_blocks is breaking things, so it need to be removed. I only comment this out now.